+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Timeout with status code 408 after five seconds.
+
+
 ## 1.1.0
 
 **2022-06-06**