+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Timeout with status code 408 after five seconds.
+- Add <kbd>r</kbd> and <kbd>n</kbd> keybinds on the paste view.
 
 
 ## 1.1.0